\name{crbr08}
\docType{data}
\alias{crbr08}
\title{The Cooper River Bridge Run 2008 Finishing Times}
\description{
  This dataset contains the finishing times for 29246 participants of the 2008 Cooper River Bridge Run. The four fields are the participants home state (US) or country, age in years, gun time in minutes, and chip time in minutes. The gun time is the interval between the official start of the race and crossing the finish line. The chip time is the interval between crossing the start and finish lines.
}
\usage{data(crbr08)}
\format{A data frame with 29246 rows and 4 columns}
\source{The Cooper River Bridge Run, 2008}
\references{
 Cooper River Bridge Run Executive Committee (2008) 
 http://www.bridgerun.com/results.html
}
\keyword{datasets}
