\name{profdpm}
\alias{profdpm-package}
\docType{package}
\title{
Profile Inference for the Dirichlet Process Mixture
}
\description{
This package performs profile inference for a class of product partition models (PPM). The Dirichlet process mixture is a specific case in this class of PPMs. This method searches for the maximum posterior (MAP) estimate for the cluster partition in the PPM.
}
\details{
\tabular{ll}{
Package: \tab profdpm\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2009-04-09\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
Contents
\itemize{
  \item{profLinear}{ Dirichlet process mixture of linear models }
  \item{profRand}{ The Rand statistic for objects of class \sQuote{profLinear} }
  \item{Rand}{ The Rand statistic }
  \item{crbr08}{ The Cooper River Bridge Run 2008 dataset }
}
}
\author{
Matt Shotwell
Maintainer: Matt Shotwell <shotwelm@musc.edu>
}

\keyword{ package }
