\name{profdpm}
\alias{profdpm-package}
\docType{package}
\title{
Profile Inference for the Dirichlet Process Mixture
}
\description{
This package performs profile inference for models utilizing the Dirichlet process mixture (DPM). This method is not based on Markov chain Monte Carlo methods, as are most other mechanisms of inference in the DPM. This methodology searches for the maximum posterior (MAP) estimate for the cluster indicators in the DPM model.
}
\details{
\tabular{ll}{
Package: \tab profdpm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-09\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
Contents
\itemize{
  \item{profLinear}{ Dirichlet process mixture of linear models }
  \item{crbr08}{ the Cooper River Bridge Run 2008 dataset }
}
}
\author{
Matt Shotwell
Maintainer: Matt Shotwell <shotwelm@musc.edu>
}

\keyword{ package }
