\name{profdpm}
\alias{profdpm-package}
\docType{package}
\title{
Profile Inference for Dirichlet Process Mixtures and Other Product Partition Models
}
\description{
This package facilitates profile inference (inference at the posterior mode) for a class of product partition models (\acronym{PPM}). The Dirichlet process mixture is a specific case in this class and is selected by default. These methods search for the maximum posterior (\acronym{MAP}) estimate for the cluster partition in the \acronym{PPM}.
}
\details{
\tabular{ll}{
Package: \tab profdpm\cr
Type: \tab Package\cr
Version: \tab 3.1\cr
Date: \tab 2012-09-07\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
Contents
\itemize{
  \item{\code{profLinear}}{ Linear Product Partition Models }
  \item{\code{profBinary}}{ Binary Product Partition Models }
  \item{\code{pci}}{ Partition Comparison Indices }
}
}
\author{
Matt Shotwell
Maintainer: Matt Shotwell <shotwelm@musc.edu>
}

\keyword{ package }
