\name{profdpm}
\alias{profdpm-package}
\docType{package}
\title{
Profile Inference for the Dirichlet Process Mixture
}
\description{
This package performs profile inference for a class of product partition models (PPM). The Dirichlet process mixture is a specific case in this class of PPMs. This method searches for the maximum posterior (MAP) estimate for the cluster partition in the PPM.
}
\details{
\tabular{ll}{
Package: \tab profdpm\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-04-15\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
Contents
\itemize{
  \item{profLinear}{ Dirichlet process mixture of linear models }
  \item{pci}{ Partition Comparison Indices }
}
}
\author{
Matt Shotwell
Maintainer: Matt Shotwell <shotwelm@musc.edu>
}

\keyword{ package }
